Refactor screens into dedicated modules and organize commons
