+++ function/__init__.py
@@ -1,9 +1,13 @@
-"""Utility helpers used throughout the application."""
+"""Utility helpers exposed at the package level."""
 
-from .database import DatabaseManager, DatabaseError, DuplicateEntryError
+"""Convenience exports for commonly used helpers."""
+
+from .cmn_database import DatabaseManager, DatabaseError, DuplicateEntryError
+from .cmn_app_state import get_app_state
 
 __all__ = [
     "DatabaseManager",
     "DatabaseError",
     "DuplicateEntryError",
+    "get_app_state",
 ]
