+++ function/cmn_resources.py
